Memory Game
This is a simple memory game that you click to flip over the card, when two cards are flipped over it checks if they are the same. If they are the same it will keep them flipped over and if they aren't the same it will flip them back. So you have to guess the location of each photo until all the photos are flipped over. Each time the cards are being shuffled into different locations so you can play the game as many times as you like.